+++ rstar/CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added `root` doc examples and traversal docs
 - Implemented `RTreeObject` for `Arc<T>` and `Rc<T>`
 
+## Fixed
+- Fix excessive memory retention in `bulk_load` from `Vec::split_off` over-capacity
+
 ## Changed
 - Made `RStar` methods take `Point` and `Envelope` as owned values where it makes sense ([PR](https://github.com/georust/rstar/pull/189))
 - Fix Clippy warning (surfaced in Rust 1.89) related to lifetime elision